Raycaster accuracy issues.
When the plane is farther than 4 meters, lets say 20 meters, it becomes 'less accurate'. Where it points is not where the click happens. It is offset from the cursor. The changes I made fixed the issue for me and might for others whom have panels that are farther than the normal 4 meters.